refactor: update public facing APIs to be more pythonic (#397)
* refactor(agents): update agent.run API

* refactor(agents): update BeeAgent API

* refactor(backend): update ChatModel.create API

* refactor(backend): update ChatModel.create_structure API

* chore: address changes in main

* refactor(emitter): update Emitter and Emitter.child APIs

---------

+++ python/beeai_framework/cancellation.py
@@ -56,7 +56,7 @@ class AbortSignal(BaseModel):
             callback()
 
     @classmethod
-    def timeout(cls, duration: int) -> "AbortSignal":
+    def timeout(cls, duration: float) -> "AbortSignal":
         signal = cls()
 
         loop = asyncio.get_event_loop()
